+++ lib/cubit_form.dart
@@ -14,3 +14,6 @@ export './src/logic/models/validations.dart';
 export './src/logic/form_cubit/form_cubit.dart';
 export 'package:flutter_bloc/flutter_bloc.dart';
 export 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
+
+export './src/logic/single_page_form/single_page_cubit.dart';
+export './src/logic/multi_page_form/multi_page_cubit.dart';